requirements.txt
asgiref==3.2.10
Django==3.1.1
django-filter==2.4.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
install==1.3.3
psycopg2-binary==2.8.6
PyJWT==1.7.1
python-dotenv==0.14.0
pytz==2020.1
sqlparse==0.3.1
